Fixed clang compiler error
Parameter pack expansion in the construct function in KokkosController was happening backwards (as a comment noted) with g++ but this behavior was not guarenteed since C++ does not guarentee the order of a functions arguments are evaluated. If the expressions to a function all consume data from a stream, you can get behavior where objects are read in the wrong order. Clang was traversing the pack in forward order causing extent_sizes to be loaded backwards with that compiler. The behavior has been adjusted to force forward order no matter the compiler and create_view has been adjusted to handle forward order accordingly. See for more: https://stackoverflow.com/questions/42047795/order-of-parameter-pack-expansion